requirements.txt
google_api_python_client==2.119.0
pandas==2.0.3
protobuf==5.27.1
pytrends==4.9.2
streamlit==1.35.0
